// apps/web/src/api-client/model/adminRideStatusHistoryItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Waymate API
 * Carpooling backend powering the Waymate web app.
 * OpenAPI spec version: 0.1.0
 */
import type { AdminRideStatusHistoryItemOldStatus } from "./adminRideStatusHistoryItemOldStatus";
import type { RideStatus } from "./rideStatus";
import type { AdminRideStatusHistoryItemChangedBy } from "./adminRideStatusHistoryItemChangedBy";

export interface AdminRideStatusHistoryItem {
    /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
    id: string;
    /** @nullable */
    oldStatus: AdminRideStatusHistoryItemOldStatus;
    newStatus: RideStatus;
    /** @nullable */
    reason: string | null;
    createdAt: string;
    /** @nullable */
    changedBy: AdminRideStatusHistoryItemChangedBy;
}